Create sketch_freewifi.ino
This sketch is a project to experiment with LilyGo T-Dongle-S3, where I have created a Wi-Fi hotspot. When a user connects to the Wi-Fi, they are redirected to the login page where they are prompted to use their Google credentials. 
If the user enters their credentials, the credentials get stored on the ESP32 and I can access them on the portal or my Serial Monitor, if I am connected to Arduino IDE.